gaze_kd_project/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
pillow>=9.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
